Update index.html
It was working in Mozilla Firefox but there was an issue in Google Chrome and Microsoft Edge (`403 Forbidden` on all the tiles).
I found [this issue](https://gitlab.com/librespacefoundation/satnogs/satnogs-network/issues/647) that looked similar and they suggested to update the library. It indeed seems to fix the issue.